URL's have changed
there are now multiple URLs for  sending commands and the commands have changed. been building and testing this for a few days now. seems solid. cant get the poll (remote status) stuff to work though for some reason. Using Fiddler on mopar.com to build out the poll URL and params...it seems i have it all right, it does a GET on the same URL as the POST... but I gave up for now.

documentation for Home Assist for the mopar sensor needs changed for this as well.